Add tests with Arquillian and modify package structure
