+++ .github/workflows/test.yaml
@@ -1,0 +1,18 @@
+name: Test
+
+on:
+  push:
+    branches: [ test ]
+
+jobs:
+  build:
+    name: Test
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: Set up JDK 1.8
+        uses: actions/setup-java@v1
+        with:
+          java-version: 1.8
+      - name: Test project with Maven
+        run: mvn -B test --file pom.xml
