+++ .idea/libraries/Maven__Arquillian_JUnit_Release.xml
@@ -1,0 +1,30 @@
+<component name="libraryTable">
+  <library name="Maven: Arquillian JUnit:Release" type="repository">
+    <properties maven-id="org.jboss.arquillian.junit:arquillian-junit-container:RELEASE" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/junit/arquillian-junit-container/1.7.0.Alpha1/arquillian-junit-container-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/junit/arquillian-junit-core/1.7.0.Alpha1/arquillian-junit-core-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/test/arquillian-test-api/1.7.0.Alpha1/arquillian-test-api-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/core/arquillian-core-api/1.7.0.Alpha1/arquillian-core-api-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/test/arquillian-test-spi/1.7.0.Alpha1/arquillian-test-spi-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/core/arquillian-core-spi/1.7.0.Alpha1/arquillian-core-spi-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/container/arquillian-container-test-api/1.7.0.Alpha1/arquillian-container-test-api-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/shrinkwrap/shrinkwrap-api/1.2.6/shrinkwrap-api-1.2.6.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/shrinkwrap/descriptors/shrinkwrap-descriptors-api-base/2.0.0/shrinkwrap-descriptors-api-base-2.0.0.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/container/arquillian-container-test-spi/1.7.0.Alpha1/arquillian-container-test-spi-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/container/arquillian-container-spi/1.7.0.Alpha1/arquillian-container-spi-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/core/arquillian-core-impl-base/1.7.0.Alpha1/arquillian-core-impl-base-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/test/arquillian-test-impl-base/1.7.0.Alpha1/arquillian-test-impl-base-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/container/arquillian-container-impl-base/1.7.0.Alpha1/arquillian-container-impl-base-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/config/arquillian-config-api/1.7.0.Alpha1/arquillian-config-api-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/config/arquillian-config-impl-base/1.7.0.Alpha1/arquillian-config-impl-base-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/config/arquillian-config-spi/1.7.0.Alpha1/arquillian-config-spi-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/shrinkwrap/descriptors/shrinkwrap-descriptors-spi/2.0.0/shrinkwrap-descriptors-spi-2.0.0.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/arquillian/container/arquillian-container-test-impl-base/1.7.0.Alpha1/arquillian-container-test-impl-base-1.7.0.Alpha1.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/shrinkwrap/shrinkwrap-impl-base/1.2.6/shrinkwrap-impl-base-1.2.6.jar!/" />
+      <root url="jar://$MAVEN_REPOSITORY$/org/jboss/shrinkwrap/shrinkwrap-spi/1.2.6/shrinkwrap-spi-1.2.6.jar!/" />
+    </CLASSES>
+    <JAVADOC />
+    <SOURCES />
+  </library>
+</component>
