--- src/main/java/Erreurs/ApiErreur.java
@@ -1,17 +1,0 @@
-package Erreurs;
-
-public class ApiErreur {
-    private String type;
-
-    public ApiErreur(String e) {
-        type = e;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-}
