--- src/main/java/Erreurs/InvalidAuthorizationException.java
@@ -1,14 +1,0 @@
-package Erreurs;
-
-public class InvalidAuthorizationException extends Exception {
-    public InvalidAuthorizationException() {
-    }
-
-    public InvalidAuthorizationException(Throwable cause) {
-        super(cause);
-    }
-
-    public InvalidAuthorizationException(String message) {
-        super(message);
-    }
-}
