--- src/main/java/Erreurs/UsernameIndisponibleErreur.java
@@ -1,18 +1,0 @@
-package Erreurs;
-
-public class UsernameIndisponibleErreur extends ApiErreur {
-    private String details;
-
-    public UsernameIndisponibleErreur(String details) {
-        super(UsernameIndisponibleErreur.class.getSimpleName());
-        this.details = details;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
-    }
-}
