--- src/main/java/Erreurs/UtilisateurMalformeErreur.java
@@ -1,18 +1,0 @@
-package Erreurs;
-
-public class UtilisateurMalformeErreur extends ApiErreur {
-    private String details;
-
-    public UtilisateurMalformeErreur(String details) {
-        super(UtilisateurMalformeErreur.class.getSimpleName());
-        this.details = details;
-    }
-
-    public String getDetails() {
-        return details;
-    }
-
-    public void setDetails(String details) {
-        this.details = details;
-    }
-}
