+++ src/main/java/martine/erreurs/ApiErreur.java
@@ -1,0 +1,17 @@
+package martine.erreurs;
+
+public class ApiErreur {
+    private String type;
+
+    public ApiErreur(String e) {
+        type = e;
+    }
+
+    public String getType() {
+        return type;
+    }
+
+    public void setType(String type) {
+        this.type = type;
+    }
+}
