+++ src/main/java/martine/erreurs/UsernameIndisponibleErreur.java
@@ -1,0 +1,18 @@
+package martine.erreurs;
+
+public class UsernameIndisponibleErreur extends ApiErreur {
+    private String details;
+
+    public UsernameIndisponibleErreur(String details) {
+        super(UsernameIndisponibleErreur.class.getSimpleName());
+        this.details = details;
+    }
+
+    public String getDetails() {
+        return details;
+    }
+
+    public void setDetails(String details) {
+        this.details = details;
+    }
+}
