+++ src/main/java/martine/erreurs/UtilisateurMalformeErreur.java
@@ -1,0 +1,18 @@
+package martine.erreurs;
+
+public class UtilisateurMalformeErreur extends ApiErreur {
+    private String details;
+
+    public UtilisateurMalformeErreur(String details) {
+        super(UtilisateurMalformeErreur.class.getSimpleName());
+        this.details = details;
+    }
+
+    public String getDetails() {
+        return details;
+    }
+
+    public void setDetails(String details) {
+        this.details = details;
+    }
+}
