+++ src/test/java/martine/api/RecetteAPITest.java
@@ -1,0 +1,41 @@
+package martine.api;
+
+import org.jboss.arquillian.container.test.api.Deployment;
+import org.jboss.arquillian.junit.Arquillian;
+import org.jboss.shrinkwrap.api.ShrinkWrap;
+import org.jboss.shrinkwrap.api.asset.EmptyAsset;
+import org.jboss.shrinkwrap.api.spec.JavaArchive;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+import javax.inject.Inject;
+
+@RunWith(Arquillian.class)
+public class RecetteAPITest {
+    @Deployment
+    public static JavaArchive createDeployment() {
+        return ShrinkWrap.create(JavaArchive.class)
+                .addPackages(true,"martine.api", "martine.erreurs", "martine.models", "martine.utils")
+                .addAsResource("META-INF/persistence.xml")
+                .addAsManifestResource(EmptyAsset.INSTANCE, "beans.xml");
+    }
+
+    @Inject
+    private RecetteAPI recetteAPI;
+
+    @Test
+    public void ajouter() {
+    }
+
+    @Test
+    public void lister() {
+    }
+
+    @Test
+    public void obtenir() {
+    }
+
+    @Test
+    public void supprimer() {
+    }
+}
