+++ src/test/resources/arquillian.xml
@@ -1,0 +1,27 @@
+<arquillian xmlns="http://jboss.org/schema/arquillian" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+            xsi:schemaLocation="
+        http://jboss.org/schema/arquillian
+        http://jboss.org/schema/arquillian/arquillian_1_0.xsd">
+
+    <!-- Sets the protocol which is how Arquillian talks and executes the tests inside the container -->
+    <defaultProtocol type="Servlet 3.0" />
+
+    <!-- Configuration to be used when the WidlFly remote profile is active -->
+    <container qualifier="widlfly-remote">
+        <configuration>
+            <property name="managementAddress">127.0.0.1</property>
+            <property name="managementPort">9990</property>
+            <property name="username">admin</property>
+            <property name="password">admin</property>
+        </configuration>
+    </container>
+
+    <!-- Configuration to be used when the WildFly managed profile is active -->
+    <container qualifier="widlfly-managed" default="true">
+        <configuration>
+            <!-- <property name="jbossHome">/Users/Mahe/Documents/work/app-web/lwildfly-19.0.0.Final</property> -->
+            <property name="jbossHome">target/wildfly-19.0.0.Final/</property>
+        </configuration>
+    </container>
+
+</arquillian>
